# autogpt_platform/frontend/src/app/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AutoGPT Agent Server",
    "summary": "AutoGPT Agent Server",
    "description": "This server is used to execute agents that are created by the AutoGPT system.",
    "version": "0.1"
  },
  "paths": {
...
    "/api/store/agents": {
      "get": {
        "tags": ["v2", "store", "public"],
        "summary": "List store agents",
        "description": "Get a paginated list of agents from the store with optional filtering and sorting.\n\nArgs:\n    featured (bool, optional): Filter to only show featured agents. Defaults to False.\n    creator (str | None, optional): Filter agents by creator username. Defaults to None.\n    sorted_by (str | None, optional): Sort agents by \"runs\" or \"rating\". Defaults to None.\n    search_query (str | None, optional): Search agents by name, subheading and description. Defaults to None.\n    category (str | None, optional): Filter agents by category. Defaults to None.\n    page (int, optional): Page number for pagination. Defaults to 1.\n    page_size (int, optional): Number of agents per page. Defaults to 20.\n\nReturns:\n    StoreAgentsResponse: Paginated list of agents matching the filters\n\nRaises:\n    HTTPException: If page or page_size are less than 1\n\nUsed for:\n- Home Page Featured Agents\n- Home Page Top Agents\n- Search Results\n- Agent Details - Other Agents By Creator\n- Agent Details - Similar Agents\n- Creator Details - Agents By Creator",
        "operationId": "getV2List store agents",
        "parameters": [
          {
            "name": "featured",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Featured"
            }
          },
          {
            "name": "creator",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Creator"
            }
          },
          {
            "name": "sorted_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Sorted By"
            }
          },
          {
            "name": "search_query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Search Query"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Category"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 1, "title": "Page" }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 20, "title": "Page Size" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StoreAgentsResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/store/agents/{username}/{agent_name}": {
      "get": {
        "tags": ["v2", "store", "public"],
        "summary": "Get specific agent",
        "description": "This is only used on the AgentDetails Page\n\nIt returns the store listing agents details.",
        "operationId": "getV2Get specific agent",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Username" }
          },
          {
            "name": "agent_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Agent Name" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StoreAgentDetails" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/store/agents/{store_listing_version_id}": {
      "get": {
        "tags": ["v2", "store"],
        "summary": "Get agent by version",
        "description": "Get Store Agent Details from Store Listing Version ID.",
        "operationId": "getV2Get agent by version",
        "parameters": [
          {
            "name": "store_listing_version_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Store Listing Version Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StoreAgentDetails" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/store/download/agents/{store_listing_version_id}": {
      "get": {
        "tags": ["v2", "store", "public"],
        "summary": "Download agent file",
        "description": "Download the agent file by streaming its content.\n\nArgs:\n    store_listing_version_id (str): The ID of the agent to download\n\nReturns:\n    StreamingResponse: A streaming response containing the agent's graph data.\n\nRaises:\n    HTTPException: If the agent is not found or an unexpected error occurs.",
        "operationId": "getV2Download agent file",
        "parameters": [
          {
            "name": "store_listing_version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The ID of the agent to download",
              "title": "Store Listing Version Id"
            },
            "description": "The ID of the agent to download"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/library/agents": {
      "get": {
        "tags": ["v2", "library", "private"],
        "summary": "List Library Agents",
        "description": "Get all agents in the user's library (both created and saved).\n\nArgs:\n    user_id: ID of the authenticated user.\n    search_term: Optional search term to filter agents by name/description.\n    filter_by: List of filters to apply (favorites, created by user).\n    sort_by: List of sorting criteria (created date, updated date).\n    page: Page number to retrieve.\n    page_size: Number of agents per page.\n\nReturns:\n    A LibraryAgentResponse containing agents and pagination metadata.\n\nRaises:\n    HTTPException: If a server/database error occurs.",
        "operationId": "getV2List library agents",
        "parameters": [
          {
            "name": "search_term",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "description": "Search term to filter agents",
              "title": "Search Term"
            },
            "description": "Search term to filter agents"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/LibraryAgentSort",
              "description": "Criteria to sort results by",
              "default": "lastExecuted"
            },
            "description": "Criteria to sort results by"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number to retrieve (must be >= 1)",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number to retrieve (must be >= 1)"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Number of agents per page (must be >= 1)",
              "default": 15,
              "title": "Page Size"
            },
            "description": "Number of agents per page (must be >= 1)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryAgentResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": { "application/json": {} }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["v2", "library", "private"],
        "summary": "Add Marketplace Agent",
        "description": "Add an agent from the marketplace to the user's library.\n\nArgs:\n    store_listing_version_id: ID of the store listing version to add.\n    user_id: ID of the authenticated user.\n\nReturns:\n    library_model.LibraryAgent: Agent added to the library\n\nRaises:\n    HTTPException(404): If the listing version is not found.\n    HTTPException(500): If a server/database error occurs.",
        "operationId": "postV2Add marketplace agent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_postV2Add_marketplace_agent"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agent added successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LibraryAgent" }
              }
            }
          },
          "404": { "description": "Store listing version not found" },
          "500": { "description": "Server error" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/library/agents/{library_agent_id}": {
      "get": {
        "tags": ["v2", "library", "private"],
        "summary": "Get Library Agent",
        "operationId": "getV2Get library agent",
        "parameters": [
          {
            "name": "library_agent_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Library Agent Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LibraryAgent" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["v2", "library", "private"],
        "summary": "Update Library Agent",
        "description": "Update the library agent with the given fields.\n\nArgs:\n    library_agent_id: ID of the library agent to update.\n    payload: Fields to update (auto_update_version, is_favorite, etc.).\n    user_id: ID of the authenticated user.\n\nRaises:\n    HTTPException(500): If a server/database error occurs.",
        "operationId": "patchV2Update library agent",
        "parameters": [
          {
            "name": "library_agent_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Library Agent Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LibraryAgentUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent updated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LibraryAgent" }
              }
            }
          },
          "500": { "description": "Server error" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["v2", "library", "private"],
        "summary": "Delete Library Agent",
        "description": "Soft-delete the specified library agent.\n\nArgs:\n    library_agent_id: ID of the library agent to delete.\n    user_id: ID of the authenticated user.\n\nReturns:\n    204 No Content if successful.\n\nRaises:\n    HTTPException(404): If the agent does not exist.\n    HTTPException(500): If a server/database error occurs.",
        "operationId": "deleteV2Delete library agent",
        "parameters": [
          {
            "name": "library_agent_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Library Agent Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "204": { "description": "Agent deleted successfully" },
          "404": { "description": "Agent not found" },
          "500": { "description": "Server error" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/library/agents/by-graph/{graph_id}": {
      "get": {
        "tags": ["v2", "library", "private"],
        "summary": "Get Library Agent By Graph Id",
        "operationId": "getV2GetLibraryAgentByGraphId",
        "parameters": [
          {
            "name": "graph_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Graph Id" }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LibraryAgent" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/library/agents/marketplace/{store_listing_version_id}": {
      "get": {
        "tags": ["v2", "library", "private", "store, library"],
        "summary": "Get Agent By Store ID",
        "description": "Get Library Agent from Store Listing Version ID.",
        "operationId": "getV2Get agent by store id",
        "parameters": [
          {
            "name": "store_listing_version_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Store Listing Version Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    { "$ref": "#/components/schemas/LibraryAgent" },
                    { "type": "null" }
                  ],
                  "title": "Response Getv2Get Agent By Store Id"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/library/agents/{library_agent_id}/fork": {
      "post": {
        "tags": ["v2", "library", "private"],
        "summary": "Fork Library Agent",
        "operationId": "postV2Fork library agent",
        "parameters": [
          {
            "name": "library_agent_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Library Agent Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LibraryAgent" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/library/agents/{library_agent_id}/setup-trigger": {
      "post": {
        "tags": ["v2", "library", "private"],
        "summary": "Setup Trigger",
        "description": "Sets up a webhook-triggered `LibraryAgentPreset` for a `LibraryAgent`.\nReturns the correspondingly created `LibraryAgentPreset` with `webhook_id` set.",
        "operationId": "postV2SetupTrigger",
        "parameters": [
          {
            "name": "library_agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ID of the library agent",
              "title": "Library Agent Id"
            },
            "description": "ID of the library agent"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggeredPresetSetupParams"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LibraryAgentPreset" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_postV2Add_marketplace_agent": {
        "properties": {
          "store_listing_version_id": {
            "type": "string",
            "title": "Store Listing Version Id"
          }
        },
        "type": "object",
        "required": ["store_listing_version_id"],
        "title": "Body_postV2Add marketplace agent"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LibraryAgentResponse": {
        "properties": {
          "agents": {
            "items": { "$ref": "#/components/schemas/LibraryAgent" },
            "type": "array",
            "title": "Agents"
          },
          "pagination": { "$ref": "#/components/schemas/Pagination" }
        },
        "type": "object",
        "required": ["agents", "pagination"],
        "title": "LibraryAgentResponse",
        "description": "Response schema for a list of library agents and pagination info."
      },
      "LibraryAgentSort": {
        "type": "string",
        "enum": ["createdAt", "updatedAt", "lastExecuted"],
        "title": "LibraryAgentSort",
        "description": "Possible sort options for sorting library agents."
      },
...
      "StoreAgentsResponse": {
        "properties": {
          "agents": {
            "items": { "$ref": "#/components/schemas/StoreAgent" },
            "type": "array",
            "title": "Agents"
          },
          "pagination": { "$ref": "#/components/schemas/Pagination" }
        },
        "type": "object",
        "required": ["agents", "pagination"],
        "title": "StoreAgentsResponse"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Postmark-Webhook-Token"
      }
    }
  }
}